+++ tests/test_switch.py
@@ -21,6 +21,56 @@ from tests.api_data import get_multiple_position_fixed_data_response
 from tests.conftest import FakeKebaKeEnergyAPI
 
 
+@pytest.mark.usefixtures("entity_registry_enabled_by_default")
+async def test_heat_circuit_switches(
+    hass: HomeAssistant,
+    config_entry: MockConfigEntry,
+    fake_api: FakeKebaKeEnergyAPI,
+) -> None:
+    fake_api.responses = [
+        MULTIPLE_POSITIONS_RESPONSE,
+        get_multiple_position_fixed_data_response(),
+        MULTIPLE_POSITIONS_DATA_RESPONSE_1,
+        # Read API after services call
+        MULTIPLE_POSITIONS_DATA_RESPONSE_1,
+    ]
+    fake_api.register_requests(config_entry.data[CONF_HOST])
+
+    await setup_integration(hass, config_entry)
+
+    heat_circuit_use_heating_curve_1: State | None = hass.states.get(
+        "switch.keba_keenergy_12345678_heat_circuit_use_heating_curve_1",
+    )
+    assert isinstance(heat_circuit_use_heating_curve_1, State)
+    assert heat_circuit_use_heating_curve_1.attributes[CONF_DEVICE_CLASS] == SwitchDeviceClass.SWITCH
+    assert heat_circuit_use_heating_curve_1.attributes[ATTR_FRIENDLY_NAME] == "Heating circuit 1 Use heating curve"
+
+
+@pytest.mark.usefixtures("entity_registry_enabled_by_default")
+async def test_heat_circuit_switches_translated(
+    hass: HomeAssistant,
+    config_entry: MockConfigEntry,
+    fake_api: FakeKebaKeEnergyAPI,
+) -> None:
+    fake_api.responses = [
+        MULTIPLE_POSITIONS_RESPONSE,
+        get_multiple_position_fixed_data_response(),
+        MULTIPLE_POSITIONS_DATA_RESPONSE_1,
+        # Read API after services call
+        MULTIPLE_POSITIONS_DATA_RESPONSE_1,
+    ]
+    fake_api.register_requests(config_entry.data[CONF_HOST])
+
+    hass.config.language = "de"
+    await setup_integration(hass, config_entry)
+
+    heat_circuit_use_heating_curve_1: State | None = hass.states.get(
+        "switch.keba_keenergy_12345678_heat_circuit_use_heating_curve_1",
+    )
+    assert isinstance(heat_circuit_use_heating_curve_1, State)
+    assert heat_circuit_use_heating_curve_1.attributes[ATTR_FRIENDLY_NAME] == "Heizkreis 1 Heizkurve verwenden"
+
+
 async def test_solar_circuit_switches(
     hass: HomeAssistant,
     config_entry: MockConfigEntry,
@@ -153,6 +203,12 @@ async def test_heat_pump_switches_translated(
             MULTIPLE_POSITIONS_DATA_RESPONSE_1,
             '[{"name": "APPL.CtrlAppl.sParam.hmiRetainData.consumer1PrioritySolar[1]", "value": "0"}, '
             '{"name": "APPL.CtrlAppl.sParam.genericHeat[2].param.priority", "value": "15"}]',
+        ),
+        (
+            "switch.keba_keenergy_12345678_heat_circuit_use_heating_curve_1",
+            SERVICE_TURN_ON,
+            MULTIPLE_POSITIONS_DATA_RESPONSE_1,
+            '[{"name": "APPL.CtrlAppl.sParam.heatCircuit[0].param.enableHeatCurveLinTab", "value": "1"}]',
         ),
     ],
 )